+++ packages/tsurlfilter/.eslintrc.cjs
@@ -11,6 +11,7 @@ module.exports = {
         'import',
         'import-newlines',
         '@typescript-eslint',
+        'jsdoc',
     ],
     extends: [
         'airbnb-base',
@@ -36,6 +37,15 @@ module.exports = {
         'no-constant-condition': ['error', { checkLoops: false }],
         '@typescript-eslint/interface-name-prefix': 'off',
         'arrow-body-style': 'off',
+
+        'jsdoc/check-tag-names': [
+            'warn',
+            {
+                // Define additional tags
+                // https://github.com/gajus/eslint-plugin-jsdoc/blob/main/docs/rules/check-tag-names.md#definedtags
+                definedTags: ['note'],
+            },
+        ],
 
         // Force proper import and export of types
         '@typescript-eslint/consistent-type-imports': [