+++ packages/tsurlfilter/src/common/error.ts
@@ -1,3 +1,6 @@
+import { ZodError } from 'zod';
+import { fromZodError } from 'zod-validation-error';
+
 type ErrorWithMessage = {
     message: string
 };
@@ -45,5 +48,10 @@ function toErrorWithMessage(maybeError: unknown): ErrorWithMessage {
  * @returns Message of the error.
  */
 export function getErrorMessage(error: unknown): string {
+    // Special case: pretty print Zod errors
+    if (error instanceof ZodError) {
+        return fromZodError(error).toString();
+    }
+
     return toErrorWithMessage(error).message;
 }