+++ packages/tsurlfilter/src/rules/declarative-converter-utils/index.ts
@@ -1,0 +1,6 @@
+export {
+    RULESET_FILE_EXT,
+    getRuleSetId,
+    getRuleSetPath,
+    extractRuleSetId,
+} from './rule-set-path';