+++ packages/tswebextension/src/lib/common/index.ts
@@ -8,6 +8,7 @@ export * from './content-script/send-app-message';
 export * from './request-type';
 export * from './error';
 export * from './constants';
+export * from './storage/core';
 export * from './storage';
 
 // TODO: used in adguard-api so maybe worth adding a new entry point for common utils