+++ packages/tswebextension/src/lib/mv3/background/index.ts
@@ -24,6 +24,7 @@ export {
     getRuleSourceText,
     type PreprocessedFilterList,
     FilterListPreprocessor,
+    preprocessedFilterListValidator,
 } from '@adguard/tsurlfilter';
 
 // re-exports to prevent collision, when both tsurlfilter and tswebextension are imported
@@ -35,6 +36,8 @@ export type {
 
 export * from './app';
 
+export { PublicFiltersApi as FiltersApi } from './public-filters-api';
+
 export { tabsApi } from '../tabs/tabs-api';
 export { TabContext } from '../tabs/tab-context';
 export { USER_FILTER_ID, MAIN_FRAME_ID } from '../../common/constants';
@@ -44,7 +47,7 @@ export type {
     SettingsConfigMV3 as SettingsConfig,
 } from './configuration';
 export { MessageType } from '../../common/message-constants';
-export { RULE_SET_NAME_PREFIX } from './filters-api';
+export { RULESET_NAME_PREFIX } from '@adguard/tsurlfilter/es/declarative-converter';
 export { companiesDbService } from '../../common/companies-db-service';
 export {
     getDomain,