+++ packages/tswebextension/test/lib/common/storage/persistent-value-container.test.ts
@@ -1,12 +1,13 @@
 import { describe, expect, it } from 'vitest';
 import browser from 'webextension-polyfill';
 
+import { BrowserStorage } from '../../../../src/lib/common/storage/core';
 import { PersistentValueContainer } from '../../../../src/lib/common/storage';
 
 describe('PersistentValueContainer', () => {
     const key = 'test-key';
     const value = 'test-value';
-    const api = browser.storage.local;
+    const api = new BrowserStorage<string>(browser.storage.local);
 
     it('should initialize the value', async () => {
         const container = new PersistentValueContainer(key, api);
@@ -26,14 +27,14 @@ describe('PersistentValueContainer', () => {
     });
 
     it('should throw an error if storage is not initialized', () => {
-        const container = new PersistentValueContainer(key, browser.storage.local);
+        const container = new PersistentValueContainer(key, new BrowserStorage<string>(browser.storage.local));
 
         expect(() => container.get()).toThrow('Storage not initialized');
         expect(() => container.set(value)).toThrow('Storage not initialized');
     });
 
     it('should throw an error if storage is already initialized', async () => {
-        const container = new PersistentValueContainer(key, browser.storage.local);
+        const container = new PersistentValueContainer(key, new BrowserStorage<string>(browser.storage.local));
         await container.init(value);
 
         await expect(container.init(value)).rejects.toThrow('Storage already initialized');