+++ packages/tswebextension/vitest.config.ts
@@ -5,7 +5,10 @@ import { MANIFEST_ENV, ManifestVersionEnv } from './tasks/constants';
 export default defineConfig({
     test: {
         environment: 'jsdom', // Enables jsdom environment for tests.
-        setupFiles: './vitest.setup.ts', // Setup file for tests.
+        setupFiles: [
+            './vitest.setup.ts',
+            'fake-indexeddb/auto',
+        ],
         exclude: [
             // node_modules are excluded by default and when we extend
             // the default config, we need to exclude them explicitly.