+++ packages/tswebextension/CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Core storages (wrappers for `chrome.storage.local` and `IndexedDB`).
+
+### Changed
+
+<!-- TODO: Check them again before release -->
+- Updated [@adguard/tsurlfilter] to `v3.2.0-alpha.0`.
+- Renamed export `RULE_SET_NAME_PREFIX` to `RULESET_NAME_PREFIX`.
+
 ### Fixed
 
 - Cosmetic rules injecting into `about:blank` iframes in MV2.